Prediction methods now live in strategies.
