+++ gpytorch/lazy/interpolated_lazy_tensor.py
@@ -8,7 +8,6 @@ from .root_lazy_tensor import RootLazyTensor
 from ..utils import sparse
 from ..utils.interpolation import left_interp, left_t_interp
 from ..utils.sparse import bdsmm
-from .. import beta_features
 
 
 class InterpolatedLazyTensor(LazyTensor):
@@ -339,157 +338,6 @@ class InterpolatedLazyTensor(LazyTensor):
             else:
                 res = res.view(batch_size, n_data, -1).sum(-1)
             return res
-
-    def exact_predictive_mean(
-        self,
-        test_train_covar,
-        full_mean,
-        train_inputs,
-        train_labels,
-        num_train,
-        likelihood,
-        precomputed_cache=None,
-        non_batch_train=False,
-    ):
-        from ..distributions import MultivariateNormal
-
-        if precomputed_cache is None:
-            train_interp_indices = self.left_interp_indices
-            train_interp_values = self.left_interp_values
-
-            train_mean = full_mean.narrow(-1, 0, train_labels.size(-1))
-
-            mvn = likelihood(MultivariateNormal(train_mean, self), train_inputs)
-            train_mean, train_train_covar = mvn.mean, mvn.lazy_covariance_matrix
-
-            train_train_covar_inv_labels = train_train_covar.inv_matmul((train_labels - train_mean).unsqueeze(-1))
-
-            # New root factor
-            base_size = self.base_lazy_tensor.size(-1)
-            precomputed_cache = self.base_lazy_tensor.matmul(
-                left_t_interp(train_interp_indices, train_interp_values, train_train_covar_inv_labels, base_size)
-            )
-
-            # Prevent backprop through this variable
-            precomputed_cache = precomputed_cache.detach()
-
-        # Compute the exact predictive posterior
-        n_test = test_train_covar.size(-2)
-        test_mean = full_mean.narrow(-1, num_train, n_test)
-        test_interp_indices = test_train_covar.left_interp_indices
-        test_interp_values = test_train_covar.left_interp_values
-        res = left_interp(test_interp_indices, test_interp_values, precomputed_cache).squeeze(-1) + test_mean
-        return res, precomputed_cache
-
-    def _exact_predictive_covar_inv_quad_form_cache(self, train_train_covar_inv_root, test_train_covar):
-        train_interp_indices = test_train_covar.right_interp_indices
-        train_interp_values = test_train_covar.right_interp_values
-        base_lazy_tensor = test_train_covar.base_lazy_tensor
-        base_size = base_lazy_tensor.size(-1)
-        res = base_lazy_tensor.matmul(
-            left_t_interp(train_interp_indices, train_interp_values, train_train_covar_inv_root, base_size)
-        )
-        return res
-
-    def _exact_predictive_covar_inv_quad_form_root(self, precomputed_cache, test_train_covar):
-        # Here the precomputed cache represents K_UU W S,
-        # where S S^T = (K_XX + sigma^2 I)^-1
-        test_interp_indices = test_train_covar.left_interp_indices
-        test_interp_values = test_train_covar.left_interp_values
-        res = left_interp(test_interp_indices, test_interp_values, precomputed_cache)
-        return res
-
-    def exact_predictive_covar(
-        self,
-        test_train_covar,
-        test_test_covar,
-        train_inputs,
-        num_train,
-        likelihood,
-        precomputed_cache=None,
-        non_batch_train=False
-    ):
-        from ..distributions import MultivariateNormal
-
-        if not beta_features.fast_pred_var.on() and not beta_features.fast_pred_samples.on():
-            return super(InterpolatedLazyTensor, self).exact_predictive_covar(
-                test_train_covar, test_test_covar, train_inputs, num_train, likelihood, precomputed_cache
-            )
-
-        train_interp_indices = self.left_interp_indices
-        train_interp_values = self.left_interp_values
-        test_interp_indices = test_train_covar.left_interp_indices
-        test_interp_values = test_train_covar.left_interp_values
-
-        if (
-            precomputed_cache is None
-            or (beta_features.fast_pred_samples.on() and precomputed_cache[0] is None)
-            or (not beta_features.fast_pred_samples.on() and precomputed_cache[1] is None)
-        ):
-            # Get inverse root
-
-            grv = MultivariateNormal(torch.zeros(1), self)
-            train_train_covar = likelihood(grv, train_inputs).lazy_covariance_matrix
-
-            # Get probe vectors for inverse root
-            num_probe_vectors = beta_features.fast_pred_var.num_probe_vectors()
-            batch_size = train_interp_indices.size(0)
-            n_inducing = self.base_lazy_tensor.size(-1)
-            vector_indices = torch.randperm(n_inducing).type_as(train_interp_indices)
-            probe_vector_indices = vector_indices[:num_probe_vectors]
-            test_vector_indices = vector_indices[num_probe_vectors : 2 * num_probe_vectors]
-
-            probe_interp_indices = probe_vector_indices.unsqueeze(1)
-            probe_test_interp_indices = test_vector_indices.unsqueeze(1)
-            probe_interp_values = torch.ones(num_probe_vectors, 1, dtype=self.dtype, device=self.device)
-            if train_interp_indices.ndimension() == 3:
-                probe_interp_indices = probe_interp_indices.unsqueeze(0).expand(batch_size, num_probe_vectors, 1)
-                probe_test_interp_indices = probe_test_interp_indices.unsqueeze(0)
-                probe_test_interp_indices = probe_test_interp_indices.expand(batch_size, num_probe_vectors, 1)
-                probe_interp_values = probe_interp_values.unsqueeze(0).expand(batch_size, num_probe_vectors, 1)
-
-            probe_vectors = InterpolatedLazyTensor(
-                self.base_lazy_tensor,
-                train_interp_indices,
-                train_interp_values,
-                probe_interp_indices,
-                probe_interp_values,
-            ).evaluate()
-            test_vectors = InterpolatedLazyTensor(
-                self.base_lazy_tensor,
-                train_interp_indices,
-                train_interp_values,
-                probe_test_interp_indices,
-                probe_interp_values,
-            ).evaluate()
-
-            # Get inverse root
-            train_train_covar_inv_root = train_train_covar.root_inv_decomposition(probe_vectors, test_vectors).root
-            train_train_covar_inv_root = train_train_covar_inv_root.evaluate()
-
-            # New root factor
-            root = self._exact_predictive_covar_inv_quad_form_cache(train_train_covar_inv_root, test_train_covar)
-
-            # Precomputed factor
-            if beta_features.fast_pred_samples.on():
-                inside = self.base_lazy_tensor + RootLazyTensor(root).mul(-1)
-                inside_root = inside.root_decomposition().root.evaluate()
-                # Prevent backprop through this variable
-                inside_root = inside_root.detach()
-                precomputed_cache = inside_root, None
-            else:
-                # Prevent backprop through this variable
-                root = root.detach()
-                precomputed_cache = None, root
-
-        # Compute the exact predictive posterior
-        if beta_features.fast_pred_samples.on():
-            res = self._exact_predictive_covar_inv_quad_form_root(precomputed_cache[0], test_train_covar)
-            res = RootLazyTensor(res)
-        else:
-            root = left_interp(test_interp_indices, test_interp_values, precomputed_cache[1])
-            res = test_test_covar + RootLazyTensor(root).mul(-1)
-        return res, precomputed_cache
 
     def matmul(self, tensor):
         # We're using a custom matmul here, because it is significantly faster than
